Add error handler in case if user does not check any checkboxes

+++ includes/delete.inc.php
@@ -2,15 +2,20 @@
 require_once '../classes/deleteProduct.class.php';
 
 if ($_SERVER["REQUEST_METHOD"] == "POST") {
-	if (isset($_POST["product_selected"]) && is_array($_POST["product_selected"])) {
-		$selectedSkus = $_POST["product_selected"];
+	if (isset($_POST['deleteProduct'])) {
+		if (isset($_POST["product_selected"]) && is_array($_POST["product_selected"])) {
+			$selectedSkus = $_POST["product_selected"];
 
-		$deleteProd = new DeleteProduct();
+			$deleteProd = new DeleteProduct();
+			$deleteProd->deleteAll($selectedSkus);
 
-		$deleteProd->deleteAll($selectedSkus);
 
-		// Redirect back to the main page or display a success message
-		header("Location: ../index.php?deletion_success=1");
-		exit();
+			// Redirect back to the main page or display a success message
+			header("Location: ../index.php?deletion_success=1");
+			exit();
+		} else {
+			header('Location: ../index.php?error=product_not_selected!');
+			exit();
+		}
 	}
 }
